+++ run_hls_fullpfalgo.tcl
@@ -7,13 +7,13 @@
 # open the project, don't forget to reset
 open_project -reset proj_fullpfalgo_15151504_240MHz_II6
 set_top pfalgo3_full
-add_files firmware/simple_fullpfalgo.cpp "-DHLS_pipeline_II=6"
+add_files firmware/simple_fullpfalgo.cpp  -cflags "-DHLS_pipeline_II=6"
 add_files -tb simple_fullpfalgo_test.cpp
 add_files -tb simple_fullpfalgo_ref.cpp
-add_files -tb pattern_serializer.cpp
-add_files -tb test_utils.cpp
+add_files -tb utils/pattern_serializer.cpp
+add_files -tb utils/test_utils.cpp
 add_files -tb DiscretePFInputs.h -cflags "-std=c++0x"
-add_files -tb DiscretePFInputs_IO.h -cflags "-std=c++0x"
+add_files -tb utils/DiscretePFInputs_IO.h -cflags "-std=c++0x"
 add_files -tb data/regions_TTbar_PU140.dump
 
 # reset the solution
