+++ run_hls_fullpfalgo_mp7.tcl
@@ -7,10 +7,10 @@ set_top ${l1pfTopFunc}
 add_files firmware/simple_fullpfalgo.cpp -cflags "-DTESTMP7 -DHLS_pipeline_II=2"
 add_files -tb simple_fullpfalgo_test.cpp  -cflags "-DTESTMP7 -DHLS_pipeline_II=2 -DMP7_TOP_FUNC=${l1pfTopFunc} -DMP7_REF_FUNC=${l1pfRefFunc} -DMP7_VALIDATE=${l1pfValidate}"
 add_files -tb simple_fullpfalgo_ref.cpp -cflags "-DTESTMP7"
-add_files -tb pattern_serializer.cpp -cflags "-DTESTMP7"
-add_files -tb test_utils.cpp -cflags "-DTESTMP7"
+add_files -tb utils/pattern_serializer.cpp -cflags "-DTESTMP7"
+add_files -tb utils/test_utils.cpp -cflags "-DTESTMP7"
 add_files -tb DiscretePFInputs.h    -cflags "-DTESTMP7 -std=c++0x"
-add_files -tb DiscretePFInputs_IO.h -cflags "-DTESTMP7 -std=c++0x"
+add_files -tb utils/DiscretePFInputs_IO.h -cflags "-DTESTMP7 -std=c++0x"
 add_files -tb data/regions_TTbar_PU140.dump
 
 # reset the solution
