+++ simple_fullpfalgo_test.cpp
@@ -1,9 +1,9 @@
 #include <cstdio>
 #include "firmware/simple_fullpfalgo.h"
-#include "random_inputs.h"
-#include "DiscretePFInputs_IO.h"
-#include "pattern_serializer.h"
-#include "test_utils.h"
+#include "utils/random_inputs.h"
+#include "utils/DiscretePFInputs_IO.h"
+#include "utils/pattern_serializer.h"
+#include "utils/test_utils.h"
 
 #define NTEST 500
