+++ utils/pattern_serializer.cpp
