+++ utils/test_utils.h
@@ -1,4 +1,4 @@
-#include "firmware/data.h"
+#include "../firmware/data.h"
 
 bool track_equals(const TkObj &out_ref, const TkObj &out, const char *what, int idx) ;
 bool had_equals(const HadCaloObj &out_ref, const HadCaloObj &out, const char *what, int idx) ;
