+++ vertexing/simple_vtx_ref.cpp
